style(Test): 인수 테스트 위치 이동
