+++ README.md
@@ -101,7 +101,7 @@ This project is [MIT](https://github.com/next-step/atdd-subway-service/blob/mast
 ### 3단계 - 인증을 통한 기능 구현
 
 1. 요구사항
-    - [ ] 토큰 발급 기능(로그인) 인수 테스트 만들기
+    - [x] 토큰 발급 기능(로그인) 인수 테스트 만들기
     - [ ] 인증 - 내 정보 조회 기능 완성하기
     - [ ] 인증 - 즐겨 찾기 기능 완성하기
 2. 목록정리
