+++ src/main/java/nextstep/subway/auth/application/AuthorizationException.java
@@ -1,10 +1,7 @@
 package nextstep.subway.auth.application;
 
 import nextstep.subway.exception.ErrorMessage;
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.ResponseStatus;
 
-@ResponseStatus(HttpStatus.UNAUTHORIZED)
 public class AuthorizationException extends RuntimeException {
 
     private final ErrorMessage errorMessage;
